API: Fix invalid index.
