+++ main/commands/all/api/rules_remove_test.go
@@ -3,7 +3,7 @@ package api
 import "testing"
 
 func TestSingleIndexFlagRejectsDuplicate(t *testing.T) {
-	flag := &singleIndexFlag{value: -1}
+	flag := &singleIndexFlag{}
 	if err := flag.Set("0"); err != nil {
 		t.Fatal(err)
 	}
@@ -17,9 +17,25 @@ func TestSingleIndexFlagRejectsDuplicate(t *testing.T) {
 
 func TestSingleIndexFlagRejectsInvalidValues(t *testing.T) {
 	for _, value := range []string{"-1", "invalid"} {
-		flag := &singleIndexFlag{value: -1}
+		flag := &singleIndexFlag{}
 		if err := flag.Set(value); err == nil {
 			t.Errorf("expected index %q to be rejected", value)
 		}
+		if flag.set {
+			t.Errorf("invalid index %q marked the flag as explicitly set", value)
+		}
+	}
+}
+
+func TestSingleIndexFlagTracksExplicitZero(t *testing.T) {
+	flag := &singleIndexFlag{}
+	if flag.set {
+		t.Fatal("new index flag must be unset")
+	}
+	if err := flag.Set("0"); err != nil {
+		t.Fatal(err)
+	}
+	if !flag.set || flag.value != 0 {
+		t.Fatalf("explicit zero was not preserved: %+v", flag)
 	}
 }
